Trabalho Breno/Calculadora/requirements.txt
grpcio>=1.51.3
grpcio-tools>=1.51.3
pybreaker>=1.0.1